requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Bioinformatics
biopython>=1.79
pysam>=0.17.0
cutadapt>=3.4

# Deep Learning
torch>=1.11.0
torch-audio>=0.11.0
transformers>=4.15.0
pytorch-lightning>=1.5.0

# Sequence Analysis
muscle>=3.8.1551
blast>=2.12.0

# Clustering
umap-learn>=0.5.3
hdbscan>=0.8.28
faiss-cpu>=1.7.2

# Visualization
streamlit>=1.8.0
plotly>=5.6.0
bokeh>=2.4.0
seaborn>=0.11.0
matplotlib>=3.5.0

# Data Processing
dask>=2022.2.0
joblib>=1.1.0
tqdm>=4.62.0

# Utilities
click>=8.0.0
pyyaml>=6.0
jsonschema>=4.0.0
requests>=2.27.0

# Security
python-magic>=0.4.27    # MIME type detection
python-magic-bin>=0.4.14  # Windows libmagic binaries

# Database
psycopg2-binary>=2.9.0  # For PostgreSQL
sqlalchemy>=1.4.0       # ORM for database abstraction

# Backup & Cloud Storage (optional)
boto3>=1.26.0           # AWS S3 backup support
azure-storage-blob>=12.14.0  # Azure Blob backup support
google-cloud-storage>=2.7.0  # Google Cloud Storage backup support

# Task Queue
celery>=5.3.4           # Distributed task queue
redis>=5.0.0            # Message broker for Celery
flower>=2.0.1           # Celery monitoring tool

# Monitoring & Observability
prometheus-client>=0.19.0  # Prometheus metrics exporter
psutil>=5.9.0           # System metrics
celery-exporter>=1.6.0  # Celery metrics for Prometheus

# Development & Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-xdist>=3.3.0       # Parallel test execution
pytest-timeout>=2.1.0     # Test timeout handling
pytest-benchmark>=4.0.0   # Performance benchmarking
faker>=20.0.0             # Fake data generation
factory-boy>=3.3.0        # Test fixture factories
hypothesis>=6.88.0        # Property-based testing
freezegun>=1.2.0          # Time mocking
responses>=0.23.0         # HTTP mocking
black>=22.0.0
flake8>=4.0.0
mypy>=0.931
isort>=5.12.0
pylint>=2.17.0
bandit[toml]>=1.7.5
safety>=2.3.0
pip-audit>=2.4.0

# Workflow Management
snakemake>=7.0.0
nextflow>=21.10.0

# Cloud & Deployment
docker>=5.0.0
kubernetes>=23.0.0
mlflow>=1.24.0

# Production Server
gunicorn>=21.2.0         # WSGI HTTP Server
gevent>=23.9.0           # Async workers for Gunicorn
setproctitle>=1.3.3      # Process naming

# Load Testing
locust>=2.20.0           # Load testing framework

# Compression
brotli>=1.1.0            # Brotli compression
zstandard>=0.22.0        # Zstandard compression